+++ packages/client/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this npm module are documented in this file.
 
 ## [10.8.0] - unreleased
 
-- Updated all `@codingame/monaco-vscode` packages to `26.0.1`
+- Updated all `@codingame/monaco-vscode` packages to `26.2.1`
+- Dropped eslint and rely fully on oxlint.
 
 ## [10.7.0] - 2026-02-04